refactor custom hooks to its own file
